+++ src/blog/Models/BlogPostCategory.cs
@@ -7,9 +7,9 @@ namespace blog.Models
 {
   public class BlogPostCategory
   {
-    public int BlogPostId { get; set; }
+    public long BlogPostId { get; set; }
     public BlogPost BlogPost { get; set; }
-    public int CategoryId { get; set; }
+    public long CategoryId { get; set; }
     public Category Category { get; set; }
   }
 }
